Add rotation
